This script create a VHDX with the Nano Server image, add a new VM on the Hyper-V host and attach the VHDX to this VM
Copy the NanoServerGenerator from a Windows Server 2016 ISO to a local
folder, then, creation of a VHDX with the image of NanoServer and add a
VM in the Hyper-V Host.

The VM used the VHDX created by NanoServerGenerator and will be
connected to the virtual network specified. By default, the VM will have
"512 Mb" of RAM.